requirements.txt
aiodns==3.5.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.9.0
appdirs==1.4.4
asyncio==3.4.3
attrs==25.3.0
beautifulsoup4==4.13.4
blinker==1.9.0
ccxt==4.4.95
certifi==2025.7.9
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
contourpy==1.3.2
cryptography==45.0.5
cycler==0.12.1
fastapi==0.116.1
Flask==3.1.1
fonttools==4.58.5
frozendict==2.4.6
frozenlist==1.7.0
h11==0.16.0
html5lib==1.1
idna==3.10
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.1
kiwisolver==1.4.8
loguru==0.7.3
lxml==6.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
mplfinance==0.12.10b0
multidict==6.6.3
multitasking==0.0.11
numpy==2.3.1
packaging==25.0
pandas==2.3.1
pandas_ta==0.3.14b0
peewee==3.18.2
pillow==11.3.0
pluggy==1.6.0
propcache==0.3.2
pyarrow==21.0.0
pycares==4.9.0
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
pyparsing==3.2.3
pytest==8.4.1
pytest-asyncio==1.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.2
requests==2.32.4
scikit-learn==1.7.0
scipy==1.16.0
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
starlette==0.47.2
threadpoolctl==3.6.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
webencodings==0.5.1
Werkzeug==3.1.3
win32_setctime==1.2.0
yarl==1.20.1
yfinance==0.2.36
